+++ src/libs/screenComponents/sware-home-component/organisms/HomeScreen.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { View, Text, StyleSheet, FlatList, ScrollView } from 'react-native';
-import FormScreenA from './FormScreenA';
+import FormScreenA from './FormScreenA/FormScreenA';
 import FormScreenJ from './FormScreenJ';
-import CardScreenA from './CardScreenA';
-import CamCardA from './CamCardA';
-import PropertiesA from './PropertiesA';
-import FlatListA from './FlatListA';
+import CardScreenA from './CardScreenA/CardScreenA';
+import CamCardA from './CamCardA/CamCardA';
+import PropertiesA from './PropertiesA/PropertiesA';
+import FlatListA from './FlatListA/FlatListA';
 import CardscreenJ from './CardscreenJ';
 import FlatlistJ from './FlatlistJ';
