--- src/libs/screenComponents/sware-home-component/organisms/PropertiesA.js
@@ -1,39 +1,0 @@
-// Card.js
-import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
-
-
-const PropertiesA = (props) => {
-  return (
-    <View style={styles.card}>
-      <Text style={styles.title}>{props.title}</Text>
-      <Text style={styles.content}>{props.content}</Text>
-    </View>
-  );
-};
-
-
-const styles = StyleSheet.create({
-  card: {
-    backgroundColor: 'white',
-    padding: 20,
-    marginVertical: 10,
-    marginHorizontal: 20,
-    borderRadius: 10,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.8,
-    shadowRadius: 2,
-    elevation: 5,
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  content: {
-    fontSize: 16,
-    marginTop: 10,
-  },
-});
-
-export default PropertiesA;
